Revert "Make V8DetectionProcessor._classesCount byte instead of ushort"
This reverts commit ac47f704045d6267ee2106aed49c77daa38e4d65.

+++ TensorWeaver.Yolo/YoloV8DetectionsProcessor.cs
@@ -24,7 +24,7 @@ public sealed class YoloV8DetectionsProcessor : OutputProcessor<List<Detection>>
 
 	public YoloV8DetectionsProcessor(YoloMetadata metadata)
 	{
-		_classesCount = (byte)metadata.ClassesNames.Length;
+		_classesCount = (ushort)metadata.ClassesNames.Length;
 		_imageSize = metadata.ImageSize;
 	}
 
@@ -59,7 +59,7 @@ public sealed class YoloV8DetectionsProcessor : OutputProcessor<List<Detection>>
 		return _suppressor.Suppress(detections);
 	}
 
-	private readonly byte _classesCount;
+	private readonly ushort _classesCount;
 	private readonly NonMaxSuppressor _suppressor = new();
 	private readonly Vector2D<int> _imageSize;
